src/tests.rs
```rust
#[cfg(test)]
mod tests {
    use super::*;
    use std::path::Path;
    use std::fs::File;
    use std::io::Read;

    #[test]
    fn test_error_handling() {
        let error = crate::error::CustomError::new("Test error");
        assert_eq!(error.to_string(), "Test error");
    }

    #[test]
    fn test_util_functions() {
        let result = crate::utils::some_util_function();
        assert!(result.is_ok());
    }

    #[test]
    fn test_config_data() {
        let config = crate::config::get_config();
        assert!(config.is_ok());
    }

    #[test]
    fn test_app_state() {
        let app = crate::app::App::new();
        assert!(app.is_ok());
    }

    #[test]
    fn test_commands() {
        let result = crate::commands::some_command();
        assert!(result.is_ok());
    }

    #[test]
    fn test_event_handlers() {
        let result = crate::events::some_event_handler();
        assert!(result.is_ok());
    }

    #[test]
    fn test_setup_functions() {
        let result = crate::setup::some_setup_function();
        assert!(result.is_ok());
    }

    #[test]
    fn test_build_scripts() {
        let result = crate::build::some_build_script();
        assert!(result.is_ok());
    }

    #[test]
    fn test_cargo_files() {
        assert!(Path::new("Cargo.toml").exists());
        assert!(Path::new("Cargo.lock").exists());
    }

    #[test]
    fn test_tauri_conf() {
        let mut file = File::open("src/tauri.conf.json").unwrap();
        let mut contents = String::new();
        file.read_to_string(&mut contents).unwrap();
        assert!(contents.contains("tauri"));
    }
}
```